+++ cmake/config_checks.cmake
@@ -36,6 +36,7 @@ set( HAVE_GTS       ${GTS_FOUND}        )
 if(with_zlib AND ZLIB_FOUND)
   set(HAVE_ZLIB 1)
 endif()
+set(HAVE_PANGOCAIRO ${PANGOCAIRO_FOUND})
 
 if(LTDL_FOUND)
   set(ENABLE_LTDL 1)