+++ tests/test_edge_node_overlap_utilities.h
@@ -43,6 +43,7 @@ struct check_options {
 struct graph_options {
   std::string_view rankdir = "TB";
   std::string_view node_shape = "polygon";
+  double node_fontsize = 14;
   double node_penwidth = 1;
   std::string_view dir = "forward";
   double edge_penwidth = 1;